requirements.txt
streamlit
evds
pandas
python-dateutil
openpyxl
plotly